+++ pytest_helm_charts/utils.py
@@ -7,7 +7,7 @@ import pykube.exceptions
 from pykube import HTTPClient
 
 from pytest_helm_charts.clusters import Cluster
-from pytest_helm_charts.errors import WaitTimeoutError
+from pytest_helm_charts.errors import WaitTimeoutError, ObjectStatusError
 
 DEFAULT_DELETE_TIMEOUT_SEC = 120
 
@@ -21,40 +21,49 @@ FactoryFunc = Callable[..., T]
 MetaFactoryFunc = Callable[[pykube.HTTPClient, List[T]], FactoryFunc]
 
 
-def wait_for_namespaced_objects_condition(
+def wait_for_objects_condition(  # noqa: C901
     kube_client: HTTPClient,
-    obj_type: Type[TNS],
+    obj_type: Type[T],
     obj_names: List[str],
-    objs_namespace: str,
-    obj_condition_func: Callable[[TNS], bool],
+    objs_namespace: Optional[str],
+    obj_condition_func: Callable[[T], bool],
     timeout_sec: int,
     missing_ok: bool,
-) -> List[TNS]:
+    failure_condition_func: Optional[Callable[[T], bool]] = None,
+) -> List[T]:
     """
-    Block until all the namespaced kubernetes objects of type `obj_type` pass `obj_condition_fun` or timeout is reached.
+    Block until all the kubernetes objects of type `obj_type` pass `obj_condition_fun` or timeout is reached.
+        If `objs_namespace` is None, objects are treated as cluster-scope, otherwise as namespace-scope. If
+        optional `failure_condition_func` is passed, it is executed when objects are refreshed and if it evaluates to
+        `True`, it throws `ObjectStatusError` exception.
 
     Args:
         kube_client: client to use to connect to the k8s cluster
         obj_type: type of the objects to check; they most be derived from
-            [NamespacedAPIObject](pykube.objects.NamespacedAPIObject)
-        obj_names: a list of object resource names to check; all of the objects must be pass `obj_condition_fun`
-            for this function to succeed
-        objs_namespace: namespace where all the resources are created (single namespace for all resources)
+            [APIObject](pykube.objects.APIObject)
+        obj_names: a list of object resource names to check; all the objects must pass `obj_condition_fun`
+            for this function to end with success
+        objs_namespace: namespace where all the resources should be present (single namespace for all resources).
+            If 'None', then it is assumed objects passed are cluster-scope
         timeout_sec: timeout for the call
         obj_condition_func: a function that gets one instance of the resource object of type `obj_type`
             and returns boolean showing whether the object meets the condition or not. The call succeeds
             only if this `obj_condition_fun` returns `True` for every object passed in `obj_names`.
-        missing_ok: when `True`, the function ignores that some of the objects listed in the `obj_names`
+        missing_ok: when `True`, the function ignores that some objects listed in the `obj_names`
             don't exist in k8s API and waits for them to show up; when `False`, an
-            [ObjectNotFound](pykube.exceptions.ObjectDoesNotExist) exception is raised.
+            [ObjectNotExist](pykube.exceptions.ObjectDoesNotExist) exception is raised.
+        failure_condition_func: if not None, then each monitored object is passed to this function. If it
+            returns `True`, the `ObjectStatusError` is raised.
 
     Returns:
-        The list of namespaced object resources with all the objects listed in `obj_names` included.
+        The list of object resources with all the objects listed in `obj_names` included in the list.
 
     Raises:
         TimeoutError: when timeout is reached.
         pykube.exceptions.ObjectDoesNotExist: when `missing_ok == False` and one of the objects
             listed in `obj_names` can't be found in k8s API
+        ObjectStatusError: when `failure_condition_func` is not None and any of the objects in `obj_names`
+            returned `True` from this function.
 
     """
     if len(obj_names) == 0:
@@ -62,20 +71,29 @@ def wait_for_namespaced_objects_condition(
 
     retries = 0
     all_ready = False
-    matching_objs: List[TNS] = []
+    matching_objs: List[T] = []
     while retries < timeout_sec:
-        response = obj_type.objects(kube_client).filter(namespace=objs_namespace)
+        response = obj_type.objects(kube_client)
+        if objs_namespace:
+            response = response.filter(namespace=objs_namespace)
         retries += 1
         matching_objs = []
         for name in obj_names:
             try:
                 obj = response.get_by_name(name)
+                failed = failure_condition_func and failure_condition_func(obj)
+                if failed:
+                    raise ObjectStatusError(
+                        f"Object's '{obj.namespace}/{obj.name}' status shows failure when waiting "
+                        f"for the object's condition to pass."
+                    )
                 matching_objs.append(obj)
             except pykube.exceptions.ObjectDoesNotExist:
                 if missing_ok:
                     pass
                 else:
                     raise
+
         all_ready = len(matching_objs) == len(obj_names) and all(obj_condition_func(obj) for obj in matching_objs)
         if all_ready:
             break